demos/spark-ducklake-integration/pyproject.toml
[project]
name = "spark-ducklake-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "duckdb==1.3.2",
    "pyspark==3.5.1",
    "python-dotenv==1.1.1",
    "loguru==0.7.3",
    "numpy==2.3.4",
    "pandas==2.3.3",
    "pyarrow==22.0.0",
]
